Fix calculations for the vmcs_host_rsp as they were incorrect
The VMCS_HOST_RSP should point towards end of the stack_contents and when switching stack in launch_vm we should push host registers on the stack so they are reserved on the new stack. The calculations were wrong right before the launch_vm. current_host_rsp should point towards the end of the General Purpose Registers. When we switch stack we point there and then we push host GPR and then we vmlaunch.